fix: 자동차 목록 생성하는 객체 분리
