+++ src/main/kotlin/racingcarWinner/Main.kt
@@ -1,7 +1,7 @@
 package racingcarWinner
 
+import racingcarWinner.core.CarList
 import racingcarWinner.core.Winner
-import racingcarWinner.core.WinnerRacing
 import racingcarWinner.ui.InputView
 import racingcarWinner.ui.OutputView
 
@@ -9,12 +9,11 @@ fun main() {
     val carNameList = InputView.inputCarNames()
     val tryNumber = InputView.inputTryNumber()
 
-    val racing = WinnerRacing()
     val winner = Winner()
 
-    val cars = racing.setCars(carNameList)
+    val cars = CarList.setCars(carNameList)
 
-    OutputView.outputResult(cars, racing, winner, tryNumber)
+    OutputView.outputResult(cars, winner, tryNumber)
 
     OutputView.outputWinner(cars, winner)
 }
