+++ src/main/kotlin/racingcarWinner/core/CarList.kt
@@ -1,0 +1,9 @@
+package racingcarWinner.core
+
+object CarList {
+
+    fun setCars(carNames: List<String>): List<Car> =
+        carNames.map { carName ->
+            Car(carName.trim())
+        }
+}
